Re-order makefile creation and add a BoardConfig
We want to set up the vendor makefile first and then set up everything
else just in case we decide we want to add content to the main vendor
makefile later. Additionally, we throw in a clean BoardConfig that is
totally empty and could be skipped, but hey, everyone seems to be doing
it so why not.